+++ src/index.js
@@ -1,8 +1,4 @@
 #!/usr/bin/env node
 
-import 'core-js/stable';
-import 'regenerator-runtime/runtime';
-
-import Engine from './lib/';
-
-export default Engine;
+require = require("esm")(module/*, options*/)
+module.exports = require("./lib/")
